# Switch config to use atomic.Pointer