This updates the cache configuration in a concurrent safe way. There are no significant performance changes when using `atomic.Pointer` to synchronize access to the configuration: `10.82 ns/op` vs. `11.58 ns/op`.

### Benchmarks

```shell
goos: darwin
goarch: amd64
pkg: github.com/kacheio/kache/pkg/cache
cpu: Intel(R) Core(TM) i5-7360U CPU @ 2.30GHz
    │ no_sync.txt │              mutex.txt               │          atomic_value.txt          │         atomic_pointer.txt          │
    │   sec/op    │   sec/op     vs base                 │   sec/op     vs base               │    sec/op     vs base               │
N-4   10.82n ± 1%   25.01n ± 2%  +130.99% (p=0.000 n=10)   11.90n ± 1%  +9.93% (p=0.000 n=10)   11.58n ± 10%  +6.97% (p=0.001 n=10)

```